pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pixiis"
version = "0.1.0"
description = "Unified game launcher & voice dashboard with controller support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Fidelis Karuri"},
]
dependencies = [
    "faster-whisper",
    "torch",
    "numpy",
    "scipy",
    "sounddevice",
    "soundfile",
    "watchdog",
]

[project.optional-dependencies]
all = [
    "PySide6",
    "pyqtdarktheme",
    "XInput-Python",
    "pydirectinput",
    "vdf",
    "pylnk3",
    "icoextract",
    "kokoro-onnx",
    "tomli_w",
]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
]

[project.scripts]
pixiis = "pixiis.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pixiis = ["../../resources/**/*"]
